requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
av==12.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deep-translator==1.11.4
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
fasttext==0.9.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
google-crc32c==1.5.0
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.10
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.93
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathos==0.3.2
pillow==10.3.0
pox==0.3.4
ppft==1.7.6.8
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
PyExecJS==1.5.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rake-nltk==1.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
setuptools==71.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.35.0
streamlit-player==0.1.5
streamlit-webrtc==0.47.7
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
translators==5.9.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
youtube-transcript-api==0.6.2
